Remove CloudControl files
And move DPL::Provider::CloudControl definition to exoscale.rb
